Defer COPY on PORT! to the "port action"
COPY was modernized as a "new generic", but that meant it would just
copy PORT!s as a context.

This means when the verb COPY was used (as it is in ODBC) intending
to dispatch to a COPY function on the port it was not running.

As a workaround until new ideas come about, this makes COPY on a PORT!
for the new generic delegate to the old generic machinery.