Switch to $ for substitutions in SQL dialect
The "tied" Sigil gives what was the most desired option for substitution
in the SQL dialect using variables and evaluations from Rebol... the $

This is not a character in the SQL spec and is used only by some small
edge features in Postgres etc.  Those can currently be accomplised with
strings, e.g. "$var" instead of $var will literally put $var in the SQL