Unify $ and @ features, with $[...] and $[<!> ...]
This makes it so that $var and $(expr) will use the parameterized query
feature, doing a ? substitution and breaking out the data.  $[expr]
remains as a way to do splicing of SQL code into the dialect, BUT it
forbids strings by default, reducing the chances of slipping arbitrary
unprocessed string data into the query.  If you try to it gives you a
warning message and asks you to use LOAD-ed material instead of strings.

If you decide to ignore it and that you really want the string, you
can use $[<!> ...]