Add capability to do Windows/Unixlike backtrace
Address Sanitizer has been helpful in getting to the bottom of most
stack-trace-needing bug situations thus far.  But a difficult bug
that can only be reproduced when running 32-bit Linux emulation, and
only on GitHub Action Workflows with that same executable, that is
*not* triggering address sanitizer reveals that it doesn't hurt to
have some extra ammunition worked out for these cases.

Also generalizes a "Crash_Of_Last_Resort()" as a generic wrapper for
places you might put an abort()