More efficient / less invasive MOLD
The code shifted to where instead of having separate dispatch for
things like MOLD from what dispatches Generics, MOLD was just built
as a generic.  This made it "less efficient" as it builds a FRAME!
for each recursion in the mold.  However, this forces more code
through a common and important path...and there's already a dispatch
and generics strategy for when you have a FRAME!.  With all code going
through the same mechanism, it means investing in that core mechanism
will improve everything (and also, be stackless implicitly...)

However, calling the MOLD at the top level was done with rebElide.
This meant it was actually doing evaluation steps to build a frame,
and looking up the name of MOLDIFY the function from a WORD!, etc.
This meant you couldn't use PROBE() or such features during a WORD!
lookup or you'd get an infinite recursion.

This takes the step of building a FRAME! manually, which isn't actually
all that difficult.  Beyond sorting out the infinite recursion, it
makes this "slow" concept of MOLD less slow across the board.