Factor out divert-mandb and capture-git-hashes
This cuts down on redundancy in the GitHub workflows by means of
using composite actions.  Redirecting mandb and setting the GIT_COMMIT
and GIT_COMMIT_SHORT were done in many workflows, and even though it's
not much code the redundancy was bothersome.

AWS credentials couldn't be factored out...but the comments could,
and that comment could include explaining it can't be factored out.

Generally tidies up some redundant commenting and points people to
the README.md.  Cleans up the spacing.