pragma for dbghelp.lib in MSVC only
The Windows stack trace facility was hacked in to try and solve a
tricky bug, but juggling the build system to figure out how to ask
for it on the command line or not isn't a priority at the moment.

MinGW is ignoring the #pragma that includes it, so the link fails.
Just limit the windows stack dump to MSVC for now