Add testing of MinGW EXEs in a Windows Runner
Historically the cross-compiled EXEs were on a Linux runner and thus
were not tested.  However, we can add a dependent job which will use
the artifact transfers to test the build products.

(I wouldn't undertake such a thing in the pre-AI era, but since they
are here to worry over the details... why not?)