Become more robust when no Filesystem/Environment
The goal is that the core runs with low dependencies...and that's proven
e.g. by the web build which has a very foreign concept from the desktop
about what things like IO are.

But if you're using %main-startup.r there are still some dependencies
on the filesystem and environment, which haven't been sorted cleanly.
The environment extension was accommodated for but not the filesystem.

This makes an EXE willing to boot with no Filesystem extension.