TINYINT UNSIGNED fix for MySQL ODBC driver
The MySQL ODBC driver has a bug in handling of TINYINT UNSIGNED when
you request the field as an Unsigned Long.

This bug doesn't seem to exist if using the MariaDB ODBC driver with
the same SQL database.  :-/

In any case, asking for TINYINT UNSIGNED as a Byte seems to work
around the issue.