Try to resolve empty binary field issue
There were a couple of problems, one of which is just that the MySQL
driver was buggy and didn't get fixed until 9.2.0:

  https://dev.mysql.com/doc/relnotes/connector-odbc/en/news-9-2-0.html

Not knowing that was a bug, I tried other suggestions from AI but
the only fix is getting the new driver.

Then, the MariaDB ODBC driver for MySQL has an issue where you can't
do a second SQLGetData() call if you find out a length is zero, you
just have to accept the zero as the final answer.