Remove outdated comment about broken CONTINUE
There was something about the bootstrap executable not being able to
handle CONTINUE from within an ELSE.  I'm confused as to why it would
have ever not worked, but it does.  Code repeated in two places so
this makes one use the workaround and the other use the continue, and
both of these work.