Fix TCCState GC problem in %mod-tcc.c
Converting the TCC natives to use the same dispatcher as other API
natives meant that they were no longer storing a handle to the TCC
state.  This meant a garbage collection would throw out that TCCState,
rendering the function pointer dead.

This adds an extra slot to the "Details" to hold the HANDLE! to the
TCCState.