Fix case-sensitive output type issue with TCC
The last time outputting TCC files to disk worked was when comparisons
were still case-insensitive.

There was no error when a match wasn't found--it just used -1 as the
output type, and TCC wasn't reporting an error on bad output types in
the API.  This adds an error message and for the moment enforces the
cases that were being used by the client code.