Use SELECT vs. field picking for e.g. args.OS_ID
The current executable is picky and only lets you do field-looking
selections on keys in blocks that are SET-WORD!

But it would be disruptive to make the args block SET-WORD! since some
of the code is expecting alternating words and strings.

This takes the easy route of going with SELECT for now, but it points
out that the idea is flaky (it might mistake a value for a key).  This
is good enough for bootstrapping for now, but the whole thing needs
systemic review.