COMPOSE of FILE! fixes for directories in unzip
COMPOSE tries to implement a bit of "file calculus" when operating on
a FILE! value, to ensure that components that are supposed to indicate
directories have slashes on them.  The goal is to avoid glomming
together a directory name and file name with no slash, or producing
duplicate slashes.

The UNZIP code was transitioned to use COMPOSE on FILE! but hadn't
reconciled with this idea.