New GHOSTABLE and UNAFRAID mechanics
Ever since the introduction of antiforms which were willing to
"vaporize", there's been a tension between higher-level operations
that want to manipulate the invisible state and the vaporization itself.

Early attempts did things like distinguish between EVAL and an
EVAL:GHOSTABLE, but this only narrowly addressed a very general problem
and didn't really look very good either.

This shifts to an approach where function references can be marked as
being GHOSTABLE, such that if they are called directly by name they
are willing to vanish.  But functions that are not declared ghostable
must use an operator (`^`) to modify the call so that it will vanish
in the case of producing a ghost.

The scope of application is only to multi-step operations, so a single
step (such as that to fulfill a function argument) will receive the
true value.  Also, when a GROUP! is used in source, the fact that it
is a multi-step operation doesn't cause it to shift into an "afraid
of ghosts mode" until a value is produced, so that `expr` and `(expr)`
will be equivalent for single steps.

Also introduces the notion of distinct EVAL modes for BLOCK! and GROUP!
suited to their treatment of GHOST!