Formalize PARAMETER_FLAG_UNRUN
A mysterious bug happened due to the implicit conversion of ACTION!
into a FRAME! when calling MAP-EACH, when it didn't have support for
ACTION! on its interface.  This caused it to try and iterate nonexistent
fields on a generator's interface instead of calling the generator.

ACTION! and FRAME! semantic equivalence should not be assumed.  So this
goes ahead and formally makes the `<unrun>` annotation a requirement
to get the coercion.