Fix NULL from generators used by MAP-EACH/FOR-EACH
For expedience, the calls to generators to produce values goes through
the API.  It uses rebLift() to dodge the issues of wanting to generate
unstable antiforms (to test for the ERROR! done signal).

The code tried to decay the cell in place, but if it was a null then
you would trigger an alert which disallows null bit patterns to be in
cells representing API handles... as the API enforces the nullptr as
representing null cells, so that C can test it for truthiness.

This works around the problem by copying the handle into the SPARE
cell before doing the decay, so all states (including null) are legal.