Resurrect Giulio's generator and read-lines tests
With the addition of official GENERATOR and YIELDER behavior, the
implementations of Giulio's generators were removed from the core and
put into test files.

These test files either never made it into %core-tests.r, or were
deliberately removed.  In any case the code hadn't worked for some
time.

This restores the tests to functionality, and bases the READ-LINES
code on UPARSE.