Sigil Test Fixes
When SIGIL! existed as its own stringlike type it worked like this:

    >> to tag! '$
    == <$>

    >> to sigil! <$>
    == $

But Sigils are now a kind of "container".  In the spirit of the system,
TO conversions should disregard the sigil in terms of what's being
captured (just as TAG! doesn't consider its angle brackets to be part
of what is "TO" converted.

So if anything, it would be:

    >> to tag! '$
    == <>

    >> to tied! <>
    == $

That is to say it the "universal translation" of the TO form would
notice there's no payload outside the Sigil.

Rather than undertake a full implementation of the concept at this
time, this hacks in a little demo in order to replace the missing
TO conversion tests...and fixes some other tests.