Fix latent "lifted type spec" entries
For a time, it was thought that seeing `return: ['true 'false]` would
be a case of using quoted things as literally recognized.  This worked
for antiforms as well, so `return: [~(a b c)~]` would return an unlifted
version of the quasiform, e.g. the splice (a b c).

This turned out to be nowhere near as useful as e.g. `return: ['word!]`
meaning you return a "quoted word".  Thus types were decorated to mean
the thing matching would have that decoration.

Quasigroups then used for literal match list: `return: [~(true false)~]`
And this makes more sense than claiming you return a literal slice.

~[]~ still matches a PACK! that's empty, so it can mean void.

Some cases were found in tests and an AI-written RegEx to find other
outliers, but it doesn't seem there were many if that RegEx is right.