Change looping over void input to return void
The "VOID-in-NULL-out" concept is a heuristic which doesn't apply in
all cases.  For example:

   >> var: null

   >> append [a b c] opt var
   == [a b c]  ; not null

Hence the idea of a loop over void input returning null isn't a
foregone conclusion--that makes it look like a break occurred, vs.
being isomorphic to (say) iterating over an empty block.

This switches it so that it's more like looping over an empty block.