Rethink "Zero" type as "None", use Option's none
C's void is weak for metaprogramming (and a function returning void
can't have any `return X;` statement, only `return;`)  To make up
for this weakness, Needful defined a type called "Zero" and a literal
for initializing the unit type called `zero` which in the C++ build
was the only value with which Zero instances could be initialized.  It
was considered undesirable to conflate an Option(T)'s none initializer
with zero.

Further thinking on this revealed that if you just called the type None
and made the only value it could be initialized with be the none for
the disengaged state for Optional(T), then that was basically a perfect
idea.  None thus becomes Option(void).  This simplified edge cases in
the code and made it overall clearer, as well as avoiding expectation
that a type called Zero could coerce itself to the integer 0.