`<next>` => NEXT, SKIP returns series, TWO etc.
There was a sort of strange aversion I had to making NEXT just a normal
combinator that returned the next series position, so I made it a TAG!
for some reason.  Don't know why.  This changes that and it is a good
replacement for old-timey skip some of the time.

It also makes SKIP return the series position after the skip.  I don't
think it's beneficial to build ELIDE into it and make it ghosting...
use ELIDE SKIP if you want that.

This also realizes that TWO, THREE, FOUR, FIVE can be defined on lists
in terms of returning SPLICE.