Bring back PROCEDURE as LAMBDA variant
PROCEDURE was considered to be misguided when trying to turn everyting
into a FUNC as a way of not having to come up with variations of METHOD

But that--was itself--misguided.  It's better to make methodness a
thing that cooperating generators can interpet out of the spec, and
keep people free to make more generators if they want.

So this brings PROCEDURE back.