Define FIRST, SECOND, THIRD etc. more simply
Before "Sea of Words", extending modules involved involved copying the
entire object's memory into a new object for all the definitions, even
if just adding one key.

It doesn't work like that at all, and EXTEND is relatively cheap on
a MODULE!.  Take advantage of that by not doing the weird variadic
tapdance with SET-WORD at top level that used to be required.