Use FENCE! in func specs for locals vs. `<local>`
One problem with `<local>` was that it put the spec analysis into a
mode where it was doing locals, and once in this mode you could not
put it back to argument gathering.  This had various traditional
problems with /LOCAL even going back to R3-Alpha where things that
were auto generating function specs had trouble needing to find if
the /LOCAL had been made or not yet.

Using a FENCE! has the advantage of lining up with the CONSTRUCT
dialect for minimally creating objects, e.g. {field1 field2 field3}
making an object with 3 fields that are null.  It means that the
locals "mode" has a clear start and end based on the fence boundary.