Bunch of random LAMBDA-ification of FUNCs
FUNC was changed to not drop out its last result, and LAMBDA could not
type check, so a lot of simple functions got RETURN added into them.

This is kind of noisy for short lambda-style things, and also adds an
extra unwanted function call and throw-processing overhead.

It's a slow process to find all the affected functions and change them
back but this goes around and does a bit of that (things that were
found during the FENCE!-for-locals work)