Kill off `<with>` tag in function specs
There were some ideas floating around about how it could be acted upon
that you were "accessing" variables from outside scope.  But those
ideas never materialized.

Instead, we can use a decoration to indicate when a variable is coming
from "outside" by assigning the dot `.` to an object in the global
scope and then naming access as `.field` -- this is kind of like
saying that functions in a module can act as "member functions" of
some object representing shared state in the module, while having
the functions themselves not be resident in that object.