Put function description in PARAMETER of RETURN OF
A function's description used to be in an object that was linked from
a frame (the object used to be where all the parameter descriptions
were stored, until strings were stored in the PARAMETER! themselves,
at which point only the function's overall description had no home.)

This solves that problem by saying that RETURN has no description of
its own--the description of the return is instead the description of
the overall function.

To mitigate the inability to break down what the individual return
types are, text strings are now allowed within the typespec block
itself, providing even more granular documentation of results.