Improvements to LiftHolder/KindHolder
There was an issue with the constructor taking const Cell running up
against the disabled copy constructor in LiftHolder that was stopping
OnStack(Element*) from being able to have the LIFT_BYTE() taken.  Using
a forwarding constructor took care of the problem and allowed the
removal of the disabling of the copy constructor.

The HeartHolder hadn't been renamed to KindHolder (now that it holds
both the Sigil and the Heart).  This updates that.