Simplify and improve the C-DEBUG-BREAK native
The C-DEBUG-BREAK native was added a long time ago, before it was
thought to be able to easily access all native definitions from the
C code by name.  Now that we can do it, it's fairly obvious that it's
better if the C-DEBUG-BREAK native just breaks directly in the eval
stepper... and retriggers the next expression.  This permits debugging
much more easily than throwing generically to the trampoline and trying
to use ticks to break on the next step!