Non-branching unstable antiform check option
It's possible to accomplish the unstable antiform check with a mask
operation and three compares, without needing a branch.

Performance testing should be applied to determine the impacts of this
branch.

Comments and organization improved a little as well, and avoids checks
that assert unstable antiforms don't have sigils, while documenting
that they don't with the impossible() macro.