Use caret to preserve unstable antiforms
This makes @ decay unstable antiforms in the API, and ^ does not decay.

It leaves ^ also performing its "as is" behavior on expressions that
can be ghosts--which dovetails nicely with the API purpose.

There were some bugs in the @ behavior which this addresses, including
that the CELL_FLAG_NOTE used for typechecking was conflicting with the
CELL_FLAG_NOTE used for indicating a quasiform was actually an antiform
in suspended animation.  Also, the first item in a feed wasn't getting
the vetting of its antiform status, so `rebValue(nullptr)` would crash.