Evaluator-compatible mechanics for GET and SET
The new idea of GET and SET is not to try and be "arbitrarily helpful"
in terms of types taken, but rather to exactly match the evaluator's
offerings for types.  Consequently there's no need for GET:ANY or
SET:ANY because the GET and SET semantics are based on the ^META
decoration Sigil on what you pass in.

Also added support for GROUP!.  BLOCK! support forthcoming...