Redo "Bedrock" natives
This reworks the bootstrap versions of NATIVE and TWEAK* that have
no typespecs and improves things in a few ways:

* there's no longer a LIB slot for TWEAK*-BEDROCK or NATIVE-BEDROCK;
  it's changed so that SYM_TWEAK_P_BEDROCK and SYM_NATIVE_BEDROCK
  are just symbols, so there's no lingering declarations in LIB
  to erase.  Once LIB(NATIVE) and LIB(TWEAK_P) get overwritten with
  the typechecked versions that's it.

* The bedrock versions use Apply_Cfunc() on the NATIVE_CFUNC() of
  the non-bedrock versions, so that all the code can be kept with
  the real one with the spec vs. factoring into another function
  call.  Looks better, and faster too.

This also adds a generic facility for having SYM_XXX which are
predefined and known to the core, but that don't get builtin Symbol*.
It seemed like that could be used for the bedrock function names but
that doesn't make a ton of since as the code is loaded anyway, and
it also made asserts hard since the strings of the loaded words
could not be checked.  But we keep the implementation anyway, as it
might be useful for other purposes.