Typecheck values "in place"
The design of intrinsics made it so that frame layout required a type
checked value to generally be located in the SPARE cell of the frame
being processed.  This requirement propagated a bit so that the high
level type checking routines required things like argument type
checking or return type checking to go ahead and put the thing they
wanted to check into the SPARE cell.

However, this created an issue where the type checking wants to do
mutations to the value (e.g. to remove quote levels to match the
quoting in the type spec rule) before the check actually gets called.
When the value was being passed in the spare, these alterations would
mutate it and have to be undone.

It's better if the value lives at another pointer, and can be refreshed
into the spare cell each time it's going to be checked.  Also, it
may be that an optimized type check never has to call any frame...not
even an intrinsic, so it's better not to pay for it.