Drop USE_FLAG_SET_WORDS_ONLY, fix TYPECHECKER
The idea that binding would have instructions like "apply to SET-WORD
only" was dodgy, especially because the modern API for asking questions
differently of WORD!s is to actually change the decorations, and if
the decorations influenced the binding that would be bad.

SPECIALIZE now works like APPLY, and does its own key lookup on each
step usually.

This fixes the non-APPLY-oriented SPECIALIZE code, as well as correctly
using the Lens for virtual binding.