Some streamlining of type checking
Type checking had been stylized in such a way that you were supposed
to put the cell to typecheck into the SPARE cell because it would
most likely be passed to an intrinsic for checking.  However, the
cell had to be morphed in order to pass it in to test removing quotes
or Sigils or un-sequencing it), meaning it was easier to just have it
in a separate cell all along and copy it back each time.

So this goes through with that change, and improves some other things.