Simplify lookback to match modern EXE
There was an idea of trying to hang onto the previous feed cell in
the evaluator, which was simplified into just putting the lookback
cell into a temporary slot in the level.

Whether that simplification should be reviewed in the modern executable
as an optimization or not is up in the air.  But the bootstrap EXE
certainly shouldn't worry about it!