Move ->value, ->gotten, ->specifier into Feed
There was an issue that was going to eventually come up, where a
subframe was made which would copy the L->value from a parent frame
into a child.  Then when variadic feeds were reified into arrays in
some GC scenarios, the value would be fixed up in one level but not
the other, leaving a stale one behind.

That motivated synchronizing with the modern EXE to put the value,
specifier, and "gotten" value into the feed.