Force_Poison_Cell() in order to DROP()
There's a protection in Poison_Cell() to try and avoid its casual
usage in places that might wipe out important information (like if
a Cell is an API cell).  This check also stops overwriting any
protected cells.

This is a good conservative check for the tool.  However, the data
stack was using it...which just made it hard to do things like put a
protected cell on the stack (which might be convenient).