Fix Level_At_End() problems (JOIN,REDUCE,ALL)
Temporarily there was a belief that you could do evaluation steps on
levels at their conceptual end, and get an "out of band" answer back
telling you it was an end.  That changed to where you are not supposed
to internally to the system ever ask that.

This broke various places, like:

    join tuple! []  ; asserted vs errored
    reduce [()]  ; asserted vs. giving you []
    all [^void ^void]  ; asserted vs. giving you void