Simplify Intrinsic_ARG() machinery
The intrinsic argument had some rules on it based on typechecking when
the typecheck could not modify the thing being checked.  This turned
out to be a bad idea--because the thing being checked had to be
mutated in various ways (removing quotes, sigils, etc.) to do each
check... so the best way to do this was to load the value into the
spare slot fresh each time.

That simplification made it easier and lighter to do the intrinsic
checking.