Fix incorrect cell corruption in debug builds
This points out a greater conceptual flaw about the corruption helper;
it was developed for the specific purpose to enable Sink() of Cells,
and cells cannot have their bits completely wiped because of format
issues...but this subtlety is not applicable to non-C++ builds.  This
means the corruption is doing something fundamentally different in
the C build.  It may be that corruption in the C build should be
disabled entirely since it lacks the necessary subtlety, but this
patches the one problematic use while thinking about that.