Hack to allow `return "halt";` in %mod-stdio.c
The stdio module was using a shorthand vs. `return rebContinue("halt");`
but it was a %sys-core.h-based extension.  This meant it didn't have
the expected VarList* that was managed and had an inherited binding
on hand.  (rebContinue() picks up LIBREBOL_BINDING_NAME() implicilty,
but without the API call macro the plain return had no way to pick
that up, and only had the native's Level* on hand...which does not
have access to the binding).

It's possible to detect this situation conservatively and just allow
such modules to use code in LIB.