Don't enforce USED()/UNUSED() on ARG()
This allows you to do INCLUDE_PARAMS_OF_XXX and not worry about whether
you use all the arguments.  There were diminishing returns to the
warning (especially considering that librebol natives would not get
the warnings) and the requirement to always exhaustively mention all
the parameters discouraged the use of INCLUDE_PARAMS_OF_XXX in places
like generics, and just generally created noise where you couldn't see
the important USED()/UNUSED() cases.

The desire to use enums for the parameter number and optionality made
it harder to preserve the warnings by conditionally marking the
typedef USED().  So it pressed the issue to where the benefits of the
warnings had to be looked at starkly.

Under the new thought, you depend on the warnings for variables that
you have extracted that are meaningful at the point you use the
include params macro... so it is that definition of variables that
becomes the "did you use it or not" check.