Series APPEND/CHANGE/INSERT code on new generics
This makes a start in "layering" the implementation of APPEND and CHANGE
and INSERT in a clearer way; by separating the higher-level decision
making from the lower-level series machinery a bit better.  It takes
advantage of the fact that at the native entry points, common work
can be done vs. repeating the work in the old generics switch()
statement.

As a small step forward, the AM_PART flag is handled at the higher
level and the lower levels can take for granted that the part length
was calculated already.  This is accomplished by some simplifications
on the higher-level of how PART is handled by CHANGE