Split branches in f-modify.c for len/size
If you are doing modifications on strands in the f-modify code, then
you always have to know the length in codepoints to do the change.
But it was fairly mixed up in terms of the branches which did have
to worry about the codepoint count vs. those that could assume that
it wasn't relevant.

This takes a step toward tidying the code up by going through a level
of dereference, and saying that the length in codepoints is actually
always behind a null pointer when it's not relevant.  (This should
be able to be cleaned up in such a way that does not cost a
dereference, but it's just to get started on improving the situation
and making any mistakes more obvious.)