Unstable antiform API handling for casts
When casts were being checked, it was exposing that now that Value* in
the API can store unstable values, some cast(Stable*, p) cases were
not truthful.

This pushes some Value* handling into places that seemed to stick out
in the pattern.  (There might be more since the unstable antiform in
API changes are recent.)