Fix emitter problem for {locals} not at end
The hack in EMIT-INCLUDE-PARAMS-MACRO for getting the {string} from
the bootstrap executable to provide the information for the FENCE!'d
locals in function specs only worked when the locals were at the end.
But locals are put in arbitrary slots to act as placeholders to
accomplish frame compatibility.

This rethinks it to just erase the { and } which gets the intended
result for now.  However, it should probably turn them into ( and )
in order to be able to make the LOCAL() accessor distinct.