Rework Contravariance To Support Wrapper Types
This adds support for wrapped types in Needful to be used with Sink()
and Init(), and generally improves the conversoin behaviors...as it
had been the case that OnStack() had a lot of composability problems.

Also renames Need() as Exact(), and doesn't make it actually a
contravariant form because it would be holding invalid bits (actual
contravariance is only applicable to the types whose data is empty
or garbage and hence why the inversion exists).  So Exact() just very
narrowly works with the type it mentions with no derived types.

Overall this strengthens checks in the system and permits better use
of wrapped or instrucmentation types.