Use Need() to stop non-optional logic ARG() checks
Historical Rebol used ARG() and REF() as two different ways of testing
slots in the frame.  REF() tested for refinements.

This was changed to where ARG() would act differently for refinements
vs. non-refinements, such that it would be testable by logic due to
a transformation done that would turn Cell* to nullptr if a nulled
bit pattern was in a refinement slot.  But this didn't come with a
safety guarantee that you'd be stopped from logically testing a
non-refinement (which would never be falsey, because the nulled cell
would not be processed for the bit pattern and would just be a truthy
pointer like anything else).

This switches over to using `Need()` for non-Option() parameters and
that offers the desired check in C++ enhanced builds.  In C builds
you just have to be careful (but before you had to be careful picking
REF() vs. ARG() and there was no guarantee you chose right).