Change "Index" concept to "Ordinal"
The idea of making Index be 1-based and Offset be 0-based didn't really
pan out in the terminology, as Index came to be the series index in
logical units while offset would be the byte offset (e.g. where a
codepoint is in UTF-8 bytes).

So the 1-based index concept is changed to "Ordinal".  This means
hesitant usages of the REBIDX type can just be Index now.