Further homogenize arguments in %f-modify.c
This moves further making the Modify_List() and Modify_String_Or_Blob()
an operation whose parameterization doesn't depend on the caller,
taking out more conditional behavior based on the "op" being either
INSERT or CHANGE.

This adds a parameter so that there's a LIMIT and a PART, instead of
making the decision down the line to split the single argument one
way or the other.  The front end of APPEND and INSERT then make the
decision to put their PART in the frame slot where CHANGE would think
of that as being a LIMIT and not a PART.

Rethinking it in this way makes it possible to encode "no limit" or
"no part" as merely leaving the refinement slots as nulled, instead
of turning them into an arbitrarily large integer (which was having
overflow issues on 32-bit builds due to a recent change).  This use
of NULL to mean UNLIMITED instead of arbitrary large numbers is a
shift that has been spreading through the system slowly over time,
and it makes things a lot more intentional.