Rethink GLOM interface and PENDING
This redoes GLOM to no longer use splices, and to also use quoted as
a lightweight proxy container--which may or not be a winner of an
idea, but it's something to think about.

Pending ceases being an input parameter and is defined as a local using
the new locals definition dialect, with a notation in source to help
cue it's a proxied output as `{pending}: [...]`.