Rudimentary SPLICE! matching in UPARSE
This leverages FIND--although FIND does not support splice matching on
BINARY! or ANY-STRING?.  So this fakes it, by running a JOIN operation.
It opens the doors to more optimal implementations in the future.

The primary purpose of this was to be able to use NONE finds in PARSE
as "always match" that works even with @RULE