Resurrect IF-returns-VOID, as IF-returns-GHOST!
Given GHOST!-safety, the forces that made it a bad idea to create
"heavy voids" before (as empty-pack-inside-empty-pack) now can work
with a notion that empty pack is itself a heavy void...with GHOST! as
the light void.

This satisfyingly brings back the ability to make ELSE and THEN react
to both light void and light null.

WHEN is rethought as the "opt in" variant that gives NONE.