Codify constness and rule match of QUOTED!
Lists would return elements out of the list as matches for QUOTED!
vs. the rule itself unquoted, which quoted matches in things like
strings could not do.

This changes to return the rule unquoted for all types, for consistency
as well as making sure the experience of rules is always const (by
making PARSE take the rules as const).