Decay intrinsic arguments before intrinsic called
Unlifted values in packs are becoming a mechanism for assigning in the
dual band, and these "dual packs" can run code when they decay (e.g.
to run the getter or setter, or to fetch the alias).

With arbitrary code running on decay, this can't be done from the
inside of an intrinsic.

This pushes the responsibility of decay onto the intrinsic caller, as
well as the "opt-out" functionality.