Crude first cut at making ALIAS assignable
This is some hacking that's just enough to get it to where rather than
having to be an infix function, ALIAS can be a function that just
returns a PACK! with an unlifted meta-word! or meta-tuple! in it, and
this is a signal that if it's assigned to a variable then that variable
becomes an alias.

It's fairly rickety and needs to be generalized, but this is enough
to make the implementation of STATIC work better.